pyproject.toml
[project]
name = "shining-quasar"
version = "0.1.0"
description = "Enterprise-grade workflow system for evaluating and rewriting Q&A responses using Google Gemini AI."
readme = "README.md"
requires-python = ">=3.10, <3.14"
dependencies = [
    "google-generativeai>=0.8.3",
    "python-dotenv>=1.0.1",
    "aiofiles>=23.2.1",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "aiolimiter>=1.1.0",
    "python-json-logger>=2.0.7",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-neo4j>=0.6.0",
    "langchain-openai>=1.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.6.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "types-aiofiles>=24.1.0.20240410",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]

[tool.mypy]
python_version = "3.13"
warn_return_any = false
warn_unused_configs = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "."
ignore_missing_imports = true
exclude = ["notion-neo4j-graph/", ".venv/", "build/"]

[tool.uv.workspace]
members = ["notion-neo4j-graph"]
